#212 [feat] 팝업화면 구현 (공포지수 제외)
