+++ src/pages/Home/Home.tsx
@@ -4,7 +4,11 @@ import { detectPWA } from '@utils/Detector';
 import CardList from '@components/CardList/CardList';
 import { ContentsItemContainer, ContentsItemContent, ContentsItemTitle } from '@components/Common/ContentsItem.Style';
 import Keywords from '@components/Keywords/Keywords';
-import PWAInfoPopUp from '@components/PopUp/PWAInfoPopUp';
+import DescentPopUp from '@components/PopUp/DescentPopUp/DescentPopUp';
+import FearPopUp from '@components/PopUp/FearPopUp/FearPopUp';
+import HotPopUp from '@components/PopUp/HotPopUp/HotPopUp';
+import PWAInfoPopUp from '@components/PopUp/PWAinfoPopUp/PWAInfoPopUp';
+import RisingPopUp from '@components/PopUp/RisingPopUp/RisingPopUp';
 import StockTable from '@components/StockTable/StockTable';
 import DownSVG from '@assets/icons/down.svg?react';
 import UpSVG from '@assets/icons/up.svg?react';
@@ -29,6 +33,12 @@ const Home = () => {
   const hotStocksApiRef = useRef({} as React.ContextType<typeof VisibilityContext>);
   const risingStocksApiRef = useRef({} as React.ContextType<typeof VisibilityContext>);
   const descentStocksApiRef = useRef({} as React.ContextType<typeof VisibilityContext>);
+
+  const [isPopupOpen, setPopupOpen] = useState([false, false, false, false]);
+
+  const togglePopup = (index: number) => {
+    setPopupOpen((prev) => prev.map((value, idx) => (idx === index ? !value : value)));
+  };
   const tabMenu = ['국내주식', '해외주식'];
 
   const stockIndices = [
@@ -72,7 +82,7 @@ const Home = () => {
               return (
                 <IndexItem key={stockIndex}>
                   <IndexInfoContainer>
-                    {stockIndex} {idx === 0 && <InfoSVG className="btn_info" onClick={() => {}} />}
+                    {stockIndex} {idx === 0 && <InfoSVG className="btn_info" onClick={() => togglePopup(0)} />}
                   </IndexInfoContainer>
                   <IndexDeltaScore delta={delta}>
                     {Math.abs(delta)}
@@ -81,42 +91,46 @@ const Home = () => {
                 </IndexItem>
               );
             })}
+            {isPopupOpen[0] && <FearPopUp onClose={() => togglePopup(0)} />}
           </IndicesContainer>
         )}
         <ContentsItemContainer>
           <ContentsItemTitle color="primary50">
             지금 가장<StyledSpan color="primary50">HOT</StyledSpan> 한
             <ZipyoSVG />
-            <InfoSVG className="btn_info" onClick={() => {}} />
+            <InfoSVG className="btn_info" onClick={() => togglePopup(1)} />
           </ContentsItemTitle>
           <ContentsItemContent>
             <CardList apiRef={hotStocksApiRef} name={'HOT'} index={tabIndex} />
           </ContentsItemContent>
           <StyledText>전일 08:24 기준</StyledText>
+          {isPopupOpen[1] && <HotPopUp onClose={() => togglePopup(1)} />}
         </ContentsItemContainer>
 
         <ContentsItemContainer>
           <ContentsItemTitle color="red">
             지금 민심 <StyledSpan color="red">떡상</StyledSpan> 중인
             <ZipyoSVG />
-            <InfoSVG className="btn_info" onClick={() => {}} />
+            <InfoSVG className="btn_info" onClick={() => togglePopup(2)} />
           </ContentsItemTitle>
           <ContentsItemContent>
             <CardList apiRef={risingStocksApiRef} name={'RISING'} index={tabIndex} />
           </ContentsItemContent>
           <StyledText>전일 08:24 기준</StyledText>
+          {isPopupOpen[2] && <RisingPopUp onClose={() => togglePopup(2)} />}
         </ContentsItemContainer>
 
         <ContentsItemContainer>
           <ContentsItemTitle color="blue">
             지금 민심 <StyledSpan color="blue">떡락</StyledSpan> 중인
             <ZipyoSVG />
-            <InfoSVG className="btn_info" onClick={() => {}} />
+            <InfoSVG className="btn_info" onClick={() => togglePopup(3)} />
           </ContentsItemTitle>
           <ContentsItemContent>
             <CardList apiRef={descentStocksApiRef} name={'DESCENT'} index={tabIndex} />
           </ContentsItemContent>
           <StyledText>전일 08:24 기준</StyledText>
+          {isPopupOpen[3] && <DescentPopUp onClose={() => togglePopup(3)} />}
         </ContentsItemContainer>
         <StockTable />
         <Keywords />
